+++ src/ValueObject/IdList.php
@@ -6,7 +6,9 @@ namespace DigitalCraftsman\Ids\ValueObject;
 
 use DigitalCraftsman\Ids\ValueObject\Exception\IdAlreadyInList;
 use DigitalCraftsman\Ids\ValueObject\Exception\IdClassNotHandledInList;
+use DigitalCraftsman\Ids\ValueObject\Exception\IdListDoesContainEveryId;
 use DigitalCraftsman\Ids\ValueObject\Exception\IdListDoesContainId;
+use DigitalCraftsman\Ids\ValueObject\Exception\IdListDoesContainNoneIds;
 use DigitalCraftsman\Ids\ValueObject\Exception\IdListDoesNotContainEveryId;
 use DigitalCraftsman\Ids\ValueObject\Exception\IdListDoesNotContainId;
 use DigitalCraftsman\Ids\ValueObject\Exception\IdListDoesNotContainSomeIds;
@@ -122,11 +124,24 @@ abstract class IdList implements \IteratorAggregate, \Countable
         return new static($ids);
     }
 
+    /** @param static $idList */
+    public function addIds(self $idList): static
+    {
+        $this->mustContainNoneIds($idList);
+
+        $newIds = $this->ids;
+        foreach ($idList as $id) {
+            $newIds[] = $id;
+        }
+
+        return new static($newIds);
+    }
+
     /** @param T $id */
     public function addIdWhenNotInList(Id $id): static
     {
         if ($this->containsId($id)) {
-            return $this;
+            return new static($this->ids);
         }
 
         $ids = $this->ids;
@@ -135,15 +150,67 @@ abstract class IdList implements \IteratorAggregate, \Countable
         return new static($ids);
     }
 
+    /** @param static $idList */
+    public function addIdsWhenNotInList(self $idList): static
+    {
+        $newIds = $this->ids;
+        foreach ($idList as $id) {
+            if ($this->notContainsId($id)) {
+                $newIds[] = $id;
+            }
+        }
+
+        return new static($newIds);
+    }
+
     /** @param T $id */
     public function removeId(Id $id): static
     {
-        if (!$this->containsId($id)) {
-            throw new IdListDoesContainId($id);
+        $this->mustContainId($id);
+
+        $idsWithoutIdToRemove = $this->ids;
+        unset($idsWithoutIdToRemove[$id->value]);
+
+        return new static($idsWithoutIdToRemove);
+    }
+
+    /** @param static $idList */
+    public function removeIds(self $idList): static
+    {
+        $this->mustContainEveryId($idList);
+
+        $ids = [];
+        foreach ($this->ids as $id) {
+            if ($idList->notContainsId($id)) {
+                $ids[] = $id;
+            }
+        }
+
+        return new static($ids);
+    }
+
+    /** @param T $id */
+    public function removeIdWhenInList(Id $id): static
+    {
+        if ($this->notContainsId($id)) {
+            return new static($this->ids);
         }
 
         $idsWithoutIdToRemove = $this->ids;
         unset($idsWithoutIdToRemove[$id->value]);
+
+        return new static($idsWithoutIdToRemove);
+    }
+
+    /** @param static $idList */
+    public function removeIdsWhenInList(self $idList): static
+    {
+        $idsWithoutIdToRemove = $this->ids;
+        foreach ($idList as $id) {
+            if (array_key_exists($id->value, $idsWithoutIdToRemove)) {
+                unset($idsWithoutIdToRemove[$id->value]);
+            }
+        }
 
         return new static($idsWithoutIdToRemove);
     }
@@ -264,6 +331,18 @@ abstract class IdList implements \IteratorAggregate, \Countable
         return true;
     }
 
+    public function notContainsEveryId(self $idList): bool
+    {
+        foreach ($idList as $id) {
+            if ($this->notContainsId($id)) {
+                return true;
+            }
+        }
+
+        return false;
+    }
+
+    /** Opposite function is not called notContainsSomeIds, but containsNoneIds */
     public function containsSomeIds(self $idList): bool
     {
         foreach ($idList as $id) {
@@ -273,6 +352,17 @@ abstract class IdList implements \IteratorAggregate, \Countable
         }
 
         return false;
+    }
+
+    public function containsNoneIds(self $idList): bool
+    {
+        foreach ($idList as $id) {
+            if ($this->containsId($id)) {
+                return false;
+            }
+        }
+
+        return true;
     }
 
     /** @param static $idList */
@@ -347,11 +437,27 @@ abstract class IdList implements \IteratorAggregate, \Countable
         }
     }
 
+    /** @throws IdListDoesContainEveryId */
+    public function mustNotContainEveryId(self $idList): void
+    {
+        if (!$this->notContainsEveryId($idList)) {
+            throw new IdListDoesContainEveryId();
+        }
+    }
+
     /** @throws IdListDoesNotContainSomeIds */
     public function mustContainSomeIds(self $idList): void
     {
         if (!$this->containsSomeIds($idList)) {
             throw new IdListDoesNotContainSomeIds();
+        }
+    }
+
+    /** @throws IdListDoesContainNoneIds */
+    public function mustContainNoneIds(self $idList): void
+    {
+        if (!$this->containsNoneIds($idList)) {
+            throw new IdListDoesContainNoneIds();
         }
     }